extract ZIP handling logic into PluginArchive utility class for reuse and simplify PluginFactory implementation
